운동 기록 흐름 정리와 랜덤 매칭

+++ app/src/main/java/com/example/capstonandroid/network/dto/Track.kt
@@ -3,7 +3,7 @@ package com.example.capstonandroid.network.dto
 data class Track(
     val __v: Int,
     val _id: String,
-    val altitude: List<Int>,
+    val altitude: List<Double>,
     val checkPoint: List<List<Double>>,
     val createdAt: String,
     val description: String,
